requirements.txt
fastapi
uvicorn
httpx
python-dotenv
paho-mqtt
supabase
pydantic
